# feat(gazelle): pure golang helper