--- a/gazelle/BUILD.bazel
+++ b/gazelle/BUILD.bazel
@@ -12,7 +12,7 @@ gazelle(
     name = "gazelle_update_repos",
     args = [
         "-from_file=go.mod",
-        "-to_macro=deps.bzl%gazelle_deps",
+        "-to_macro=deps.bzl%go_deps",
         "-prune",
     ],
     command = "update-repos",