--- a/gazelle/MODULE.bazel
+++ b/gazelle/MODULE.bazel
@@ -20,6 +20,7 @@ use_repo(
     "com_github_smacker_go_tree_sitter",
     "com_github_stretchr_testify",
     "in_gopkg_yaml_v2",
+    "org_golang_x_sync",
 )
 
 non_module_deps = use_extension("//:def.bzl", "non_module_deps")