--- a/gazelle/MODULE.bazel
+++ b/gazelle/MODULE.bazel
@@ -4,7 +4,7 @@ module(
     compatibility_level = 1,
 )
 
-bazel_dep(name = "bazel_skylib", version = "1.5.0")
+bazel_dep(name = "bazel_skylib", version = "1.6.1")
 bazel_dep(name = "rules_python", version = "0.18.0")
 bazel_dep(name = "rules_go", version = "0.41.0", repo_name = "io_bazel_rules_go")
 bazel_dep(name = "gazelle", version = "0.33.0", repo_name = "bazel_gazelle")
@@ -24,4 +24,4 @@ use_repo(
 )
 
 non_module_deps = use_extension("//:def.bzl", "non_module_deps")
-use_repo(non_module_deps, "python_stdlib_list")
+use_repo(non_module_deps, "python_stdlib_list_3_11")