--- a/gazelle/MODULE.bazel
+++ b/gazelle/MODULE.bazel
@@ -17,5 +17,10 @@ use_repo(
     "com_github_bmatcuk_doublestar_v4",
     "com_github_emirpasic_gods",
     "com_github_ghodss_yaml",
+    "com_github_smacker_go_tree_sitter",
+    "com_github_stretchr_testify",
     "in_gopkg_yaml_v2",
 )
+
+non_module_deps = use_extension("//:def.bzl", "non_module_deps")
+use_repo(non_module_deps, "python_stdlib_list")