--- a/gazelle/MODULE.bazel
+++ b/gazelle/MODULE.bazel
@@ -9,6 +9,11 @@ bazel_dep(name = "rules_python", version = "0.18.0")
 bazel_dep(name = "rules_go", version = "0.41.0", repo_name = "io_bazel_rules_go")
 bazel_dep(name = "gazelle", version = "0.33.0", repo_name = "bazel_gazelle")
 
+local_path_override(
+    module_name = "rules_python",
+    path = "..",
+)
+
 go_deps = use_extension("@bazel_gazelle//:extensions.bzl", "go_deps")
 go_deps.from_file(go_mod = "//:go.mod")
 use_repo(
@@ -24,4 +29,11 @@ use_repo(
 )
 
 non_module_deps = use_extension("//:def.bzl", "non_module_deps")
-use_repo(non_module_deps, "python_stdlib_list_3_11")
+use_repo(
+    non_module_deps,
+    "python_stdlib_list_3_10",
+    "python_stdlib_list_3_11",
+    "python_stdlib_list_3_12",
+    "python_stdlib_list_3_8",
+    "python_stdlib_list_3_9",
+)