--- a/gazelle/MODULE.bazel
+++ b/gazelle/MODULE.bazel
@@ -28,9 +28,9 @@ use_repo(
     "org_golang_x_sync",
 )
 
-non_module_deps = use_extension("//:def.bzl", "non_module_deps")
+python_stdlib_list = use_extension("//python:extensions.bzl", "python_stdlib_list")
 use_repo(
-    non_module_deps,
+    python_stdlib_list,
     "python_stdlib_list_3_10",
     "python_stdlib_list_3_11",
     "python_stdlib_list_3_12",