--- a/gazelle/WORKSPACE
+++ b/gazelle/WORKSPACE
@@ -1,6 +1,6 @@
 workspace(name = "rules_python_gazelle_plugin")
 
-load("@bazel_tools//tools/build_defs/repo:http.bzl", "http_archive")
+load("@bazel_tools//tools/build_defs/repo:http.bzl", "http_archive", "http_file")
 
 http_archive(
     name = "io_bazel_rules_go",
@@ -20,6 +20,12 @@ http_archive(
     ],
 )
 
+http_file(
+    name = "python_stdlib_list",
+    sha256 = "3c1dbf991b17178d6ed3772f4fa8f64302feaf9c3385fef328a0c7ab736a79b1",
+    url = "https://raw.githubusercontent.com/pypi/stdlib-list/8cbc2067a4a0f9eee57fb541e4cd7727724b7db4/stdlib_list/lists/3.11.txt",
+)
+
 load("@bazel_gazelle//:deps.bzl", "gazelle_dependencies")
 load("@io_bazel_rules_go//go:deps.bzl", "go_register_toolchains", "go_rules_dependencies")
 
@@ -34,14 +40,9 @@ local_repository(
     path = "..",
 )
 
-load("@rules_python//python:repositories.bzl", "py_repositories", "python_register_toolchains")
+load("@rules_python//python:repositories.bzl", "py_repositories")
 
 py_repositories()
-
-python_register_toolchains(
-    name = "python_3_11",
-    python_version = "3.11",
-)
 
 load("//:deps.bzl", _py_gazelle_deps = "gazelle_deps")