--- a/gazelle/deps.bzl
+++ b/gazelle/deps.bzl
@@ -18,12 +18,25 @@ load(
     "@bazel_gazelle//:deps.bzl",
     _go_repository = "go_repository",
 )
+load("@bazel_tools//tools/build_defs/repo:http.bzl", "http_file")
 
 def go_repository(name, **kwargs):
     if name not in native.existing_rules():
         _go_repository(name = name, **kwargs)
 
+def python_stdlib_list_deps():
+    http_file(
+        name = "python_stdlib_list_3_11",
+        sha256 = "3c1dbf991b17178d6ed3772f4fa8f64302feaf9c3385fef328a0c7ab736a79b1",
+        url = "https://raw.githubusercontent.com/pypi/stdlib-list/8cbc2067a4a0f9eee57fb541e4cd7727724b7db4/stdlib_list/lists/3.11.txt",
+        downloaded_file_path = "3.11.txt",
+    )
+
 def gazelle_deps():
+    go_deps()
+    python_stdlib_list_deps()
+
+def go_deps():
     "Fetch go dependencies"
     go_repository(
         name = "co_honnef_go_tools",