--- a/gazelle/python/extensions.bzl
+++ b/gazelle/python/extensions.bzl
@@ -1,0 +1,5 @@
+"python_stdlib_list module extension for use with bzlmod"
+
+load("//python/private:extensions.bzl", _python_stdlib_list = "python_stdlib_list")
+
+python_stdlib_list = _python_stdlib_list