… CHANGELOG.md
… gazelle/go.sum
… gazelle/python/__main__.py
… gazelle/python/lifecycle.go
… gazelle/python/parse.py
… gazelle/python/parse_test.py
… gazelle/python/python_test.go
… gazelle/python/resolve.go
… gazelle/python/std_modules.go
… gazelle/python/std_modules.py
… gazelle/python/std_modules_test.go