@hunshcn 2024-05-14
now //gazelle have no python dep. So I try to remove rules_python. But it will case test error. https://buildkite.com/bazel/rules-python-python/builds/7651#018f760c-1662-459b-a5b0-2c6b72446b62
Although it's not a big impact, it puzzles me. Can someone answer it for me?