@aignas 2024-05-15
@dougthor42, @linzhp, could you test this PR in your repos? Does this speed things up? Does it slow things down? Since you two have bigger repos, it would be nice to use them as tests to ensure there are no hidden edge cases.