@hunshcn 2024-05-16
Because #1887 introduced an error (see https://github.com/bazelbuild/rules_python/pull/1905 for details). I dropped #1887 in the current branch, otherwise it may affect the test of @linzhp  (if it is a multi-language mixed repo). 

dougthor42‘s test was not affected, because I hadn't rebase main at that time. ...

It has no effect on the current pr, just a description.